requirements.txt
langchain
openai
huggingface_hub
streamlit==1.37.1
langchain==0.2.14
langchain-openai==0.1.22
python-dotenv==1.0.1
